requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
google-generativeai>=0.3.0
chromadb>=0.4.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
httpx>=0.24.0
tiktoken>=0.5.0